requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.2.0  
colorama==0.4.6
contourpy==1.1.0
cycler==0.11.0
DeepImageSearch==2.5       
Django==4.2.3
djangorestframework==3.14.0
faiss-cpu==1.7.4
filelock==3.12.2
fonttools==4.41.1
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.3
matplotlib==3.7.2
mpmath==1.3.0
networkx==3.1
numpy==1.25.1
packaging==23.1
pandas==2.0.3
Pillow==10.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
safetensors==0.3.1
six==1.16.0
sqlparse==0.4.4
sympy==1.12
timm==0.9.2
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4